Update GameScreen.kt
Cenetered the tic tac toe board buttons, set their colour and added a placeholder icon till someone clicks on it to make it either an x or o. Commented code for x and o icons. Need to discuss and implement a turn and turn time indicator which will implement later.